Fix all compiler warnings for clean cross-platform build
- Fix sequence-point undefined behavior in union initializers (Color.h, float.h)
- Fix switch case range warnings by casting to unsigned char (calc.cpp)
- Fix format specifier mismatches using PRId64/PRIu64/PRIx64 macros
- Fix multi-char constants with explicit bit shifts (inputrecord.cpp)
- Fix cdecl attribute for x86_64 (attributes.h)
- Add matching operator new to RefCounted class (ref_ptr.h)
- Fix unused return value warnings with explicit variable capture
- Fix buffer overflow/truncation issues by increasing buffer sizes
- Fix real bug: free of modified pointer after getintval (cfgfile.cpp)
- Fix EASTL strict-aliasing warnings using intermediate variables
- Add [[maybe_unused]] for intentionally unused static variables
- Use cross-platform inttypes.h format macros throughout

+++ libs/uae_lib/custom.cpp
@@ -11968,7 +11968,7 @@ static bool framewait(void)
 
 #ifdef DEBUGGER
 		if (0 || (log_vsync & 2)) {
-			write_log (_T("%06lld %06lld/%06lld %03lld%%\n"), t, vsynctimeperline, vsynctimebase, t * 100 / vsynctimebase);
+			write_log (_T("%06" PRId64 " %06" PRId64 "/%06" PRId64 " %03" PRId64 "%%\n"), (int64_t)t, (int64_t)vsynctimeperline, (int64_t)vsynctimebase, (int64_t)(t * 100 / vsynctimebase));
 		}
 #endif
 
@@ -12032,7 +12032,7 @@ static bool framewait(void)
 		vsyncmaxtime = curr_time + max;
 
 		if (1)
-			write_log (_T("%06lld:%06lld/%06lld %d %d\n"), adjust, vsynctimeperline, vstb, (int)max, maxvpos_display);
+			write_log (_T("%06" PRId64 ":%06" PRId64 "/%06" PRId64 " %d %d\n"), (int64_t)adjust, (int64_t)vsynctimeperline, (int64_t)vstb, (int)max, maxvpos_display);
 	
 	} else {
